Create add-blacklist
I am a core team member for the Unilayer project and this website was setup to scam users "to let them claim the GAIA airdrop" it asks users to connect their MM wallet and provide their seed. Please blacklist this website to prevent people from falling victim to this scam.